requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
pydantic==2.3.0
python-multipart==0.0.6
google-generativeai==0.3.1
pymupdf==1.23.5
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
cryptography==41.0.3
alembic==1.12.0
pytest==7.4.2
pytest-asyncio==0.21.1
